requirements.txt
kivy
Pillow
keyboard
